Move login buttons to page folder
