+++ app/Login/Login.tsx
@@ -3,7 +3,7 @@
 
 import { useUser } from "@/context/userContext";
 import Link from "next/link";
-import Spinner from "../Spinner/Spinner";
+import Spinner from "../../components/Spinner/Spinner";
 
 const LoginButtons: React.FC<{}> = () => {
   const { user, loading } = useUser();
